+++ aten/src/ATen/native/nested/cuda/NestedTensorTransformerFunctions.cpp
@@ -11,6 +11,7 @@
 
 #include <ATen/native/nested/NestedTensorTransformerFunctions.h>
 #include <ATen/native/nested/NestedTensorMath.h>
+#include <ATen/native/nested/NestedTensorUtils.h>
 
 namespace at {
 namespace native {
